feat: route TensorsHelper matmul row-dot through WidenPrimitives (lifts AutoDiff)
The generic MultiplyRowScalar inner dot now goes through WidenPrimitives.Dot<T>,
so BFloat16/Half matmul (owned by Linear/Attention/Conv) accelerates when the
toggle + length gate are satisfied. NarrowFloatKernels.Dot returns the narrowed
T (BFloat16/Half) instead of a raw float so the result type matches the
T[] assignment; the WidenPrimitives.Dot<T> box-unbox cast then works.
Float/double typed kernels and int/other T fall through to TensorPrimitives.Dot<T>
unchanged.

+++ src/Nivara/Primitives/NarrowFloatKernels.cs
@@ -100,11 +100,11 @@ internal static class NarrowFloatKernels
     // Dot product (matmul hot path)
     // ─────────────────────────────────────────────────────────────────────────
 
-    internal static float Dot(ReadOnlySpan<BFloat16> a, ReadOnlySpan<BFloat16> b)
-        => Vector128.IsHardwareAccelerated ? DotBf16Core(a, b) : DotScalar(a, b);
+    internal static BFloat16 Dot(ReadOnlySpan<BFloat16> a, ReadOnlySpan<BFloat16> b)
+        => (BFloat16)(Vector128.IsHardwareAccelerated ? DotBf16Core(a, b) : DotScalar(a, b));
 
-    internal static float Dot(ReadOnlySpan<Half> a, ReadOnlySpan<Half> b)
-        => Vector128.IsHardwareAccelerated ? DotHalfCore(a, b) : DotScalar(a, b);
+    internal static Half Dot(ReadOnlySpan<Half> a, ReadOnlySpan<Half> b)
+        => (Half)(Vector128.IsHardwareAccelerated ? DotHalfCore(a, b) : DotScalar(a, b));
 
     private static float DotBf16Core(ReadOnlySpan<BFloat16> a, ReadOnlySpan<BFloat16> b)
     {
